requirements.txt
# LLM Gateway v2.0 Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.3

# Caching
aiosqlite>=0.19.0
numpy>=1.26.0

# Security
python-multipart>=0.0.6

# Monitoring
prometheus-client>=0.19.0

# Anthropic
anthropic>=0.18.0

# Optional: Local Embeddings
# sentence-transformers>=2.2.0
